+++ features/fixtures/app/module/src/main/java/com/example/android/gradle/plugin/FooApp.java
@@ -1,0 +1,14 @@
+package com.example.android.gradle.plugin;
+
+import android.app.Application;
+
+import com.Bar;
+
+public class FooApp extends Application {
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        new Bar().doSomething();
+    }
+}