+++ features/scripts/bundle_one_flavor.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
+set -e
 
-cd features/fixtures/app
-./gradlew :module:clean :module:bundleFoo --stacktrace
+cd $APP_FIXTURE_DIR
+echo "Test fixture used: $APP_FIXTURE_DIR, AGP=$AGP_VERSION, Gradle=$GRADLE_WRAPPER_VERSION"
+./gradlew :module:clean :module:bundleFoo -x lint --stacktrace