+++ features/scripts/bundle_project_module.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -e
 
 # Alter git remote origin url in case protocol differs
 PREV_REMOTE=`git config --get remote.origin.url`
@@ -10,5 +11,6 @@ function resetGitConfig {
 
 trap resetGitConfig EXIT
 
-cd features/fixtures/app
-./gradlew :module:clean :module:bundle --stacktrace
+cd $APP_FIXTURE_DIR
+echo "Test fixture used: $APP_FIXTURE_DIR, AGP=$AGP_VERSION, Gradle=$GRADLE_WRAPPER_VERSION"
+./gradlew :module:clean :module:bundle -x lint --stacktrace